application-generator/src/content/python/modules/blogs/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
blinker==1.4
click==8.0.4
colorama==0.4.4
cryptography==39.0.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mysql-connector-python==8.0.29
PyMySQL==1.0.2
python-dotenv==0.19.2
python-slugify==6.1.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3